+++ src/main/java/faang/school/projectservice/controller/resource/ResourceController.java
@@ -2,7 +2,7 @@ package faang.school.projectservice.controller.resource;
 
 import com.amazonaws.services.s3.model.S3Object;
 import faang.school.projectservice.model.dto.ResourceDto;
-import faang.school.projectservice.service.resource.ResourceService;
+import faang.school.projectservice.service.ResourceService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
