+++ src/main/java/faang/school/projectservice/filter/impl/InternshipStatusFilter.java
@@ -1,5 +1,6 @@
-package faang.school.projectservice.filter.internship;
+package faang.school.projectservice.filter.impl;
 
+import faang.school.projectservice.filter.InternshipFilter;
 import faang.school.projectservice.model.dto.InternshipFilterDto;
 import faang.school.projectservice.model.entity.Internship;
 import org.springframework.stereotype.Component;
@@ -7,16 +8,14 @@ import org.springframework.stereotype.Component;
 import java.util.stream.Stream;
 
 @Component
-public class InternshipTeamRoleFilter implements InternshipFilter {
+public class InternshipStatusFilter implements InternshipFilter {
     @Override
     public boolean isApplicable(InternshipFilterDto internshipFilterDto) {
-        return internshipFilterDto.getRolePattern() != null;
+        return internshipFilterDto.getStatusPattern() != null;
     }
 
     @Override
     public Stream<Internship> apply(Stream<Internship> internshipStream, InternshipFilterDto internshipFilterDto) {
-        return internshipStream
-                .filter(internship -> internship.getInterns().stream()
-                        .anyMatch(intern -> intern.getRoles().contains(internshipFilterDto.getRolePattern())));
+        return internshipStream.filter(internship -> internship.getStatus() == internshipFilterDto.getStatusPattern());
     }
 }
