+++ src/main/java/faang/school/projectservice/filter/impl/InternshipTeamRoleFilter.java
@@ -1,5 +1,6 @@
-package faang.school.projectservice.filter.internship;
+package faang.school.projectservice.filter.impl;
 
+import faang.school.projectservice.filter.InternshipFilter;
 import faang.school.projectservice.model.dto.InternshipFilterDto;
 import faang.school.projectservice.model.entity.Internship;
 import org.springframework.stereotype.Component;
@@ -7,14 +8,16 @@ import org.springframework.stereotype.Component;
 import java.util.stream.Stream;
 
 @Component
-public class InternshipStatusFilter implements InternshipFilter {
+public class InternshipTeamRoleFilter implements InternshipFilter {
     @Override
     public boolean isApplicable(InternshipFilterDto internshipFilterDto) {
-        return internshipFilterDto.getStatusPattern() != null;
+        return internshipFilterDto.getRolePattern() != null;
     }
 
     @Override
     public Stream<Internship> apply(Stream<Internship> internshipStream, InternshipFilterDto internshipFilterDto) {
-        return internshipStream.filter(internship -> internship.getStatus() == internshipFilterDto.getStatusPattern());
+        return internshipStream
+                .filter(internship -> internship.getInterns().stream()
+                        .anyMatch(intern -> intern.getRoles().contains(internshipFilterDto.getRolePattern())));
     }
 }
