+++ src/main/java/faang/school/projectservice/filter/impl/SubProjectStatusFilter.java
@@ -1,5 +1,6 @@
-package faang.school.projectservice.service.subproject.filters;
+package faang.school.projectservice.filter.impl;
 
+import faang.school.projectservice.filter.SubProjectFilter;
 import faang.school.projectservice.model.dto.ProjectDto;
 import faang.school.projectservice.model.entity.Project;
 import org.springframework.stereotype.Component;
