+++ src/test/java/faang/school/projectservice/controller/ProjectControllerTest.java
@@ -9,7 +9,7 @@ import static org.mockito.Mockito.when;
 
 import faang.school.projectservice.controller.project.ProjectController;
 import faang.school.projectservice.dto.project.ProjectDto;
-import faang.school.projectservice.service.project.ProjectService;
+import faang.school.projectservice.service.ProjectService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
