+++ src/test/java/faang/school/projectservice/controller/resource/ResourceControllerTest.java
@@ -3,7 +3,7 @@ package faang.school.projectservice.controller.resource;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.S3Object;
 import faang.school.projectservice.model.dto.ResourceDto;
-import faang.school.projectservice.service.resource.ResourceService;
+import faang.school.projectservice.service.ResourceService;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
