+++ src/test/java/faang/school/projectservice/service/StageServiceTest.java
@@ -10,7 +10,6 @@ import faang.school.projectservice.model.entity.Task;
 import faang.school.projectservice.model.entity.Stage;
 import faang.school.projectservice.model.enums.TasksAfterDelete;
 import faang.school.projectservice.repository.StageRepository;
-import faang.school.projectservice.service.project.ProjectService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
