+++ src/test/java/faang/school/projectservice/service/resource/S3ServiceTest.java
@@ -6,6 +6,7 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
 import faang.school.projectservice.model.entity.Resource;
 import faang.school.projectservice.model.enums.ResourceStatus;
+import faang.school.projectservice.service.S3Service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
