+++ src/test/java/faang/school/projectservice/service/subproject/SubProjectServiceTest.java
@@ -7,9 +7,10 @@ import faang.school.projectservice.model.entity.Project;
 import faang.school.projectservice.model.enums.ProjectStatus;
 import faang.school.projectservice.model.enums.ProjectVisibility;
 import faang.school.projectservice.repository.ProjectRepository;
-import faang.school.projectservice.service.subproject.filters.SubProjectFilter;
-import faang.school.projectservice.service.subproject.filters.SubProjectNameFilter;
-import faang.school.projectservice.service.subproject.filters.SubProjectStatusFilter;
+import faang.school.projectservice.filter.SubProjectFilter;
+import faang.school.projectservice.filter.impl.SubProjectNameFilter;
+import faang.school.projectservice.filter.impl.SubProjectStatusFilter;
+import faang.school.projectservice.service.SubProjectService;
 import faang.school.projectservice.validator.subproject.ValidatorService;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
